The cli should provide a blocking status check.
Why: Because running a CI will require a way to run a script that
exits with either a 0 (pass) or 1 (fail).

How: The scripts works by polling an api endpoint on bitHound's
server. It assumes a few things -- mainly that you've set up the
BITHOUND_API environment variable to point to https://www.bithound.io.

No timeout is included because this command is meant to be used in
a CI (which typically impose timeouts on their executed scripts).

In the case of a failure, a message will be written to stderr.